aider: Removed redundant loop to display each file only once in the file list table.
# Aider chat conversation:

USER: When uploading a file, multiple lines are shown for each file. like this: "File Name	File Size	EXIF Date	Status	Progress
IMG_1176.jpeg	1363662 bytes		OK
IMG_1176.jpeg 1363662 bytes OK "

ASSISTANT: To address the issue of multiple lines being shown for each file upon upload, we need to modify the `app/templates/upload.html` file to ensure that each file is listed only once in the file list table. It seems that there might be a duplication in the template rendering logic.

Here is the diff for the necessary changes:

```diff
--- app/templates/upload.html
+++ app/templates/upload.html
@@ ... @@
     <div id="file-list-server">
         <table>
             <thead>
                 <tr>
                     <th>File Name</th>
                     <th>File Size</th>
                     <th>EXIF Date</th>
                     <th>Status</th>
                     <th>Progress</th>
                 </tr>
             </thead>
             <tbody>
-                {% for file in file_list %}
-                    <tr>
-                        <td>{{ file.name }}</td>
-                        <td>{{ file.size }} bytes</td>
-                        <td>{{ file.exif_date }}</td>
-                        <td>{{ file.status }}</td>
-                        <td><progress value="{{ file.progress }}" max="100"></progress></td>
-                    </tr>
-                {% endfor %}
             </tbody>
         </table>
     </div>
             <tbody>
-        <ul>
-            {% for file in file_list %}
-                <tr>
-                    <td>{{ file.name }}</td>
-                    <td>{{ file.size }} bytes</td>
-                    <td>{{ file.exif_date }}</td>
-                    <td>{{ file.status }}</td>
-                    <td><progress value="{{ file.progress }}" max="100"></progress></td>
-                </tr>
-            {% endfor %}
             </tbody>
         </table>
     </div>
```

This change removes the redundant loop that was causing each file to be listed twice. Now, each file should be listed only once in the file list table.